+++ src/mapreduce/Mapper.java
@@ -1,9 +1,9 @@
 package mapreduce;
 
 import io.Context;
-import io.IntWritable;
+import io.LongWritable;
 import io.Text;
 
 public abstract class Mapper {
-    public abstract void map(IntWritable key, Text value, Context context);
+    public abstract void map(LongWritable key, Text value, Context context);
 }
